Rename the plugin, update docstrings
